+++ app/src/main/java/com/twitterdev/rdio/app/MemoryCache.java
@@ -10,13 +10,18 @@ public class MemoryCache {
     private Map<String, SoftReference<Bitmap>> cache=Collections.synchronizedMap(new HashMap<String, SoftReference<Bitmap>>());
 
     public Bitmap get(String id){
-        if(!cache.containsKey(id))
-            return null;
         SoftReference<Bitmap> ref=cache.get(id);
-        return ref.get();
+        if(ref==null)
+            return null;
+        Bitmap bitmap=ref.get();
+        if(bitmap==null)
+            cache.remove(id);
+        return bitmap;
     }
 
     public void put(String id, Bitmap bitmap){
+        if(id==null || bitmap==null)
+            return;
         cache.put(id, new SoftReference<Bitmap>(bitmap));
     }